[Fix] Rename typing and enhance collect_env (#2420)
