+++ mmaction/utils/__init__.py
@@ -4,7 +4,7 @@ from .gradcam_utils import GradCAM
 from .misc import (VideoWriter, frame_extract, get_random_string, get_shm_dir,
                    get_thread_id)
 from .setup_env import register_all_modules
-from .typing import *  # noqa: F401,F403
+from .typing_utils import *  # noqa: F401,F403
 
 __all__ = [
     'collect_env', 'get_random_string', 'get_thread_id', 'get_shm_dir',
