+++ mmaction/utils/collect_env.py
@@ -1,4 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+import mmcv
 from mmengine.utils import get_git_hash
 from mmengine.utils.dl_utils import collect_env as collect_basic_env
 
@@ -10,6 +11,20 @@ def collect_env():
     env_info = collect_basic_env()
     env_info['MMAction2'] = (
         mmaction.__version__ + '+' + get_git_hash(digits=7))
+    env_info['MMCV'] = (mmcv.__version__)
+
+    try:
+        import mmdet
+        env_info['MMDetection'] = (mmdet.__version__)
+    except ImportError:
+        pass
+
+    try:
+        import mmpose
+        env_info['MMPose'] = (mmpose.__version__)
+    except ImportError:
+        pass
+
     return env_info
